+++ CHANGELOG.md
@@ -67,6 +67,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Handbook
     - `Subjects::Widget` component to gallery
     - `Files::Widget` component to gallery
+    - `validateList` validator
 
 ### Changed
 - Models